+++ Changes.md
@@ -1,8 +1,8 @@
 ﻿Version Information
 ===================
 
-Version 404.0.2 - In development
---------------------------------
+Version 404.0.2 - 30/05/2024
+----------------------------
 1. Fix section zero with a summary is shown when there are no modules.
 2. Add flexible modules.
 3. Only show a section if it has modules that are visible to the user when not editing.